finished json to ufo.
added working example for deleting, loading and getting a font & compile to ttf
