+++ src/data/data_utils/font_to_ufo.py
@@ -1,6 +1,5 @@
 from defcon import Font
 import extractor
-from ufo2ft import compileOTF
 
 def ttf_to_ufo(ttf_file_path: str, ufo_file_path: str):
 
@@ -26,12 +25,3 @@ def var_ttf_to_ufo(ttf_file_path: str, ufo_file_path: str):
 
     ufo.save(ufo_file_path)
 
-def ufo_to_ttf(ufo_file_path, ttf_file_path):
-    ufo = Font(ufo_file_path)
-    otf = compileOTF(ufo)
-    otf.save(ttf_file_path)
-
-if "__main__" == __name__:
-
-    ufo_to_ttf("../../../data/processed/fonts/UFO/Alef/Alef-Regular.ufo", "Alef-Regular.ttf")
-    #ufo_to_ttf("Alef-Regular.ufo", "Alef-Regular.ttf")
